WikiPageConfig::getConfigData should return an empty array when loading failed
This codepath previously threw an exception, which
breaks others parts of the code. Throwing an exception makes
_both_ get() and has() throw an exception, while returning
an empty array means has() finishes nicely (with a false),
while get still throws an exception (as calling get with has()
returning false is unexpected). That behavior is necessary for
GrowthExperimentsMultiConfig (which is a wrapper around MultiConfig)
to work. When has() returns false, MultiConfig consults the fallback(s),
but with an exception thrown, it stops processing, and the exception
propagates up to the user.

We want GrowthExperimentsMultiConfig to silently use the defaults
derived from extension.json and server config rather than
breaking the wiki.

If this happens, it will be still logged as an error, and will likely
need someone with +staff or +sysop to fix the error.

Bug: T274520
Change-Id: Ie1a17cd01a2c291556488824680232be58189e45