refreshLinkRecommendations.php: use explicit begin
While on the DB level a COMMIT does not require a BEGIN, our PHP
framework throws an excpetion when commit() is called and the
transaction session ID is not set. In theory this should not be
possible as any DB write triggers beginIfImplied(), but somehow
it is happening anyway. An explicit begin() will hopefully make
these errors go away.

Follows up Idbf3bedd82e4479c228709ecc9d5d1dd8ade104a which was
the previous attempt to fix this error.

Bug: T277702
Change-Id: Idd03ee317d8b4faf8b2401c5457d1af8dff0ccbd