Provide capability for A/B testing task types, step 2
Hide the link-recommendation task type when link
recommendations are disabled for the user and the links
task type when they are enabled. Since ConfigurationLoader
is needed in a ResourceLoader callback where the user
identity is not available, and since TaskSuggester
takes an explicit user that might be different from the
context user, don't try to filter inside those, but
instead remove the invalid task types when generating the
list of task types to pass to TaskSuggester (which happens
in NewcomerTasksUserOptionsLookup), and on the client side
where needed.

No attempts are made to filter task types in calls which
the user shouldn't be able to get to by normal means if
the respective task type is disabled:
- in the link-recommendation-specific REST APIs
- in the growthtasks API (as calls from the frontend code
  always specify the list of task types to use)

Bug: T278123
Change-Id: I18b968dae1b5dd92e3e43af130e9984df9463d30