Make variants independent from the homepage
Instead of a JS variable, set the variant via user options (this
was already mostly the case but not for the default variant)
plus ResourceLoaderGetConfigVars, so it can be used outside of
the homepage. It is assumed that the variant won't depend on
anything but the user.

The user preference and configuration variable are not renamed,
on the assumption that eventually we'll want a multidimensional
variant system and it will be more convenient to do the rename
when adding that.

Also change Utils.getUserVariant() to work the same way as
UserExperimentManager when the user option is invalid, and
expose a JS variant getter/setter for testing convenience.

Change-Id: Ifee5c88f8684f0d3da8bbef89f014e06ecade612