Disable Submit button during structured edits with no changes
Disable the save button if the user did not accept or reject any
recommendations. Previously, if a link was accepted then unaccepted
(or rejected/unrejected), the button stayed active.

This is a departure from normal VisualEditor behavior, where you
can do and undo a change and still be able to save (presumably
because it's hard to check if the document changed), but it's
implied in the Add Link specification, seems more intuitive, and
avoids having to specialcase all-skipped edits on the server side.

Bug: T269657
Change-Id: I49b78a4410b8bd8b5af7d8c1c45d9ad9277a62dc