AddLink: Remove edit source tab if present
This is very much a hack, but it will be removed when we implement the
edit mode toggle (T269653).

There might be a cleaner way to prevent Edit Source tab from appearing
to begin with but AFAIK it would involve faking user preferences to
specify that the user prefers VE or modifying VE's config set up to
insist on single tab setup.

Bug: T281594
Change-Id: I548bdbffb60ddef5f8f4341c636d5d6c66f42f16