refreshLinkRecommendations.php: Use per-wiki locks
Named locks are global. With the link recommendation service now
having plenty of capacity to spare, use per-wiki locks so the
script can run on all wikis in parallel. (We'll probably want to
parallelize even more but that requires some thought on how to
avoid the script getting into conflict with itself, while this
much should be a no-brainer.)

We might need to re-think this when there are many wikis, but
with only four initial target wikis it should be fine.

Also replaces some deprecated DB constants.

Change-Id: I68b8fbb502999c282bdaa37cda85ab33669efb2d