Add MentorDashboardDataProvider to be used by mentee overview
This adds several data providers:

== UncachedMenteeOverviewDataProvider
This provider directly calculates the required numbers using
vslow database replica (as the queries are expected to be slow).

This provider is NOT supposed to be used in UI-facing code, instead
it will be used in a maintenance script, populating the caching table. For
that reason, no caching is implemented .

== DatabaseMenteeOverviewDataProvider
This provider talks to growthexperiments_mentee_data, and
caches the result for an hour (as repeated queries will be made,
for any filtering made by the user).

== StaticMenteeOverviewDataProvider
Static version of the data provider. I intend to use
it in frontend experiments, but it will be useful in tests
as well, I think.

== updateMenteeData.php script
A maintenance script, which uses UncachedMenteeOverviewDataProvider
and MentorStore to get data about all mentees that meet the business
rules, and updates growthexperiments_mentee_data to have data only
about those mentees.

Bug: T280007
Change-Id: Idb1013303f4d045ee311a428cdbf040325fc2259