Add Link: Select visual diff mode on save dialog setup
Make sure the save dialog always opens in visual mode.

This also prevents the error that occurred when the user had no
stored preference for the diff mode, in which case VE defaults
to the diff mode matching the surface mode, but that got confused
by our custom surface mode.

Bug: T281924
Change-Id: I380c7f3b434669f5e9e11057c034885d235c87e3