Fix diff mode override
Setting the button did not fully work for some reason so instead
(temporarily) set the user preference which will be used in
ve.ui.MWSaveDialog.getSetupProcess to set the button. This seems
to avoid the issues and is also more flexible (if the user has a
preexisting preference, we could easily honor it, although the
current implementation does not).

Bug: T281924
Change-Id: I9d404c28b97231f7390c929a8a281047de6b90ae