Add Link: refine exclusion rules for finding link text matches
Try to match the logic of the link recommendation service in
excluding parts of the document that are in templates etc.
The service relies on mwparserfromhell and searches the nodes
returned by `wikicode.filter_text(recursive=False)` for matches,
which will only return text nodes which are on the top level of
the syntax tree generated by mwparserfromhell.

The Javascript version added here does a tree traversal and
excludes everything known to be generated by Parsoid (other than
<section> and <p> which do not have corresponding nodes in the
wikitext AST). A more aggressive alternative would be to explicitly
restrict the search to elements on the /p or /section/p path; we
take the more conservative approach here.

No attempt is made in this patch to handle the cases where Parsoid
turns plain text into something else (other than <p>-wrapping),
such as turning raw HTML entities into mw:Entity tags.

Bug: T267694
Change-Id: I83724e0ed15161630aa7193c32b41bc2aa3322b4
Depends-On: I6565426e9bf7084b1a33d39b2882639a03210460