Add Link: Keep focus on the link inspector dialog
Ensure focus is returned to the dialog after actions that affect
the surface (normally changing the selected annotation would leave
the focus on the surface). This makes sure the Esc key is sent to
the right place, and improves accessibility in general (e.g. makes
it possible to tab through the action buttons).

Bug: T282179
Change-Id: I5c46030d54e1d4ca2496e03b1f801c1ff4a1b056