Add Link: Suppress the blue dot on the edit button
Add Link edits aren't initiated manually so the call to attention
for the edit link doesn't make sense.

This hardcodes the task type ID which is not super elegant but it's
already used in the same file elsewhere. In the future, we might
want to have per-tasktype settings for the preferred editor
(T282800) and disable the dot when the preferred editor is
machineSuggestions.

Bug: T283094
Change-Id: Ib86d405ef70f2c4d44168bf885e7d51cb15a6598