Make it possible to deploy welcomesurvey to % of users that's not divisible by 10
Add a new key, `percentage`, to WelcomeSurveyExperimentalGroups,
and use that (if available) to determine which group to show to the user.

This is done for English Wikipedia deployment, as we want to give
WS only to 2 % of users.

Bug: T284127
Change-Id: I03b380ba40d0b41ee92088072f310f5a65d3acb5
(cherry picked from commit bf3f47da80bb332a58feea3f6ad5074b08117c55)