Prevent adding "add links" edit tag when edit is made external to interface
Tracking of "newcomer task" edits used a mechanism to cache the page IDs
the user interacted with from Special:Homepage; and then if the page ID
is in the user's cache when they've made an edit to the page, we'd
credit "newcomer tasks" for driving that edit activity. This worked for
unstructured tasks like copyediting, references, etc where the user
would be pretty unlikely to find their way back to the same article from
outside the Special:Homepage interface.

With add link, we should be a bit more precise about this, because the
user can only see the add link interface when they first arrive on the
article after clicking on a task from Special:Homepage. After they're
suggested edit session has expired (e.g. moved to a different page)
they'll get the regular VE / wikitext editing interface.

Since we have a hook that tells us is the add link interface was used in
the edit, use that to untrack pages from the user's edit tracker when we
see that the editing interface was not add link.

This does not run in a deferred update so that we execute before the
onRecentChanges_save hook, but the code being executed should add a very
minimal delay to the VisualEditor hook execution time.

Bug: T284217
Bug: T284516
Change-Id: I5bacaf1d9460c418a99734c3046880e6ee9c67e3