WelcomeSurvey: Set exp2_target_specialpage as the only group
WelcomeSurveyExperimentalGroups is overriden in WMF production,
but merge strategies can merge the config in unexpected ways (see T284599).

For this variable, it's better to set its default
to something close to the desired production config.

Right now, we don't do any A/B testing with WelcomeSurvey,
and exp2_target_specialpage is the only used group.

This patch is prompted by T284599, but does not address
the issue in any way, so not connecting it to that task.

Change-Id: I473ba04009d13206ce1e1d6d441b42651df6269a