AddLink: Ensure revision has suggestions before saving
In between the time a user clicks on an article with link suggestions,
and the time they attempt to publish an edit, other edits might have
occurred to the same article. If this occurs we want to inform the user
and take them back to suggested edits.

This patch uses VisualEditor's presave hook to check if the link
recommendation store contains content for the latest revision of the
page, and if not, it throws an error. That error is then handled in an
override of VE's saveErrorHookAborted method, which will hide the
default dialog and show an alert dialog.

Depends-On: I6b81ea318f52ec47661086d85b5cc242a3fcd0e4
Depends-On: I6eb5b5f6c42cea54adb5b6cdc919be4138e6265a
Bug: T283109
Bug: T287103
Change-Id: Ic8225933c9fa2724fb86e00277bfd3d2c74b2961