SuggestedEdits: Only log task impression for EditCardWidget
Other card widgets have their own impression
type ('se-task-pseudo-impression'), so this code should be conditioned
on having an edit card widget. That in turn means that this.logCardData
won't fail when it ends up expecting an actual task; 'task' is undefined
when looking at the end of the queue, or in case of an error.

Bug: T283546
Change-Id: I0719b33c54d2ee9d6458426cb1b2ae3b9e2df874
(cherry picked from commit 500f22b0f239964b4d33bb3a4673c408688217f8)