WikiPageConfigWriter: Apply autopatrolled rights properly
We don't do this in the other features that save edits. It usually
does not cause any issues, as the features are used by newcomers
who usually don't have autopatrolled flag. But with WikiPageConfigWriter,
things are different, and this will get noticed, sooner or later. Prevent
it by applying autopatrolled rights properly.

Bug: T285756
Change-Id: Ib25d06983ed9fbecfd79056f8c1ebc409d0ed402