selenium: Use Special:CreateAccount for account creation
This would have caused test failures when geEnabled was added in the
form descriptor. One can verify this by removing the default -1 from the
formDescriptor for geEnabled.

Note that this test may be unstable until T299193 is solved. But since
we retry tests 3 times, it is unlikely to cause a build to fail.

Bug: T313884
Bug: T292293
Change-Id: I6eeb63d895bb75a55e6ae5bab70deec01351d98c