Add a link: No-op onAnnotationClicked when the handler is called when it's added
With mw.hook, if an event is fired, and a handler added afterwards, the handler is called right away with the last given argument. When the user navigates to read and comes back, the handler is added again so it could be called if the user clicked on the annotation before navigating away.

Bug: T287610
Change-Id: Ia75102468cadee6d5bca7ae80aebfdc478bc5a7b