Decouple mentorship from the rest of the features
We need to deploy mentorship separately from
the rest of the Growth features on the English Wikipedia;
mentorship needs to be offered to less users, until we
figure out how to scale it well.

This patch introduces GEMentorshipNewAccountEnablePercentage
global, which can be optionally used to deploy mentorship
to some portion of the homepage-enabled users.

GEMentorshipNewAccountEnablePercentage is "a percentage of a percentage":
it will be only considered if user is picked for (some) Growth
features in the first place.

The default is "mentorship enabled". This has several advantages:

* users who enroll manually will be able to use all Growth features
* old users will see no change
* wikis where GEMentorshipNewAccountEnablePercentage is not used
  won't have preferences cluttered with "yes, enabled".

On enwiki, this will cause a lot of "not enabled" entries though,
IMO that's fine, for the (relatively short) duration of the test.

Bug: T287903
Change-Id: I57da73acaec8bc886f1ad42b072c9481a86ee603