Split out the link recommendation fetch/validate/store logic
...from refreshLinkRecommendations.php to a dedicated class so it
can be reused in other scripts.
The new class is kept agnostic of the task/topic configuration
hacks that are used in the maintenance script to search by ORES
topics and for filtering out task candidates which are already
tasks, so in theory it could be reused in any other place, e.g.
in a job to refresh link recommendations after a page was edited.

Also drop the SearchEngineFactory reference as it wasn't used
anywhere.

Bug: T284551
Change-Id: I452c4c26c12efd052ec04412caf0a3451615c9a9