Homepage: Remove Start module and task modules
The module and its task modules haven't been used in some time;
if we decide we need it later we can always revert this patch.

Also make StartEditing extend BaseModule so we can do away with
BaseTaskModule class.

EventLogging for StartEditing now relies on activation state of
SuggestedEdits as a proxy for whether StartEditing is "complete" or
"incomplete".

Bug: T289109
Change-Id: I2a033fd16cf6a426b972d1d44087d1fd3adff993