Mentor dashboard: Fix base font size
Because I was running an older version of Vector,
I did not notice that the fontsize for .vector-body changed
to approximately 14px, which is what we want here.

This resulted in screens like https://phabricator.wikimedia.org/F34611741
for developers who run the correct version (or use beta for testing).

Fix this by removing the override.

Bug: T278920
Change-Id: Id381673153a07b2d058135412074ef9c348b1268