Mentee overview: Fix border-radius
It appears border-radius does not work with
border-collapse: collapsed. To get the border
rounded, use border-spacing: 0 instead of border-collapse.

Also, apply overflow: hidden to crop the backround color
of the table header.

The other solution that I tried was to wrap the table
in a div, and apply the border to wrapping div.

Considering this change is smaller and appears to work,
I'm pushing it instead. However, I'm not sure if any of the approaches
comes with unintended consequences.

Bug: T278971
Change-Id: Ic79c712ffa2edea0961044cac0100a1f591a9276