NewcomerTasks: Excluded templates / categories per task type
Rework the functionality to allow for excluding templates/categories
from association with a task type so that it operates on a per task type
level instead of across all task types.

It appears to be currently unused on any on-wiki configuration so this should be
safe to do.

Update Special:EditGrowthConfig so that machine suggestion
tasks can be edited via Special:EditGrowthConfig. That's needed so that
admins can specify excluded templates and categories for image
recommendation tasks.

Output the task-type in the on-wiki JSON config. For template
based tasks it's inferred, but better to be explicit about it. That
field is not exposed as an editable option on Special:EditGrowthConfig.

Also implement validation of the structure of the 'templates',
'excludedTemplates', and 'excludedCategories' fields.

Bug: T289216
Change-Id: Id38cb56d8a345e6ddf84f325f46893f26433466d