Add image recommendations to task type A/B filter
* For users who are in the 'imagerecommendation' variant,
  enable image recommendations and make them the default task type
  selection.
* For users who are not in the variant, hide the task type, and
  remove it from the user's current selection if present.

As we currently can't handle multiple variants at the same time,
the patch won't work well when link recommendations and image
recommendations are both A/B tested (more specifically, there
wouldn't be a way for anyone to see both at the same time).
So we'll either need to remove the link recommendation logic
before setting the 'imagerecommendation' variant to any users,
or solve T288022.

Bug: T290403
Change-Id: Ie4d0ef0495b40abeff825d83bf1c28303e81c848