SuggestedEdits: Handle taskPreviewData error
In my local set up, I have a CirrusSearch backend error due to not
populating my search index. On the server side, SuggestedEdits.php
populates the taskPreviewData array with an error key and a value
containing the error message. If we don't check for the error, then
taskPreviewData is an empty object, and our code assumes that this is a
suggested edit card with no data; eventually we end up with a JavaScript
error when logCardData sees that no task is defined.

Change-Id: I29785ec8029d8bb50c2912b3b12b3a11c8012bc0