Mentor dashboard: Refactor checks a bit
Using userCanExecute() for checks like "dashboard is enabled"
is a bit of an abuse, and PermisisonError isn't perfect description
of the error.

This uses similar concept to what SpecialHomepage::handleDisabledPreference
uses.

It also adds a check that ensures mentor list is enabled,
as the dashboard fatalerrors without that list present.

In production environment, the dashboard won't be useful with such
list, so it doesn't make sense to allow it to operate without it.

After all, by default, with no list, no one is considered a mentor,
so the furthest someone would get is a permission error.

Bug: T278920
Change-Id: I9f07e2e2a3238c20d1263ec5de9beffa112b19cb