Mentor dashboard: Try to hide no JS notice from JS users
Mentee overview is a JS only feature. To inform no-JS users about that,
the server-side part includes a no-JS notice, which is later replaced
with the interface by the client.

To reduce the time for which the notice is seen by users it isn't relevant
for, apply a class to the notice, and set it to display: none when .client-js
is applied.

This makes the notice unnoticeable at localhost (except when I intentionally
disable JS), so let's see how it will work at beta.

Bug: T289901
Change-Id: I621890b186c281fe49155a35a89144482eca896a