MentorManager: Mentor list methods should not throw WikiConfigException for missing mentor list
Missing mentor list is a valid configuration (in fact, it's
the most common one as of now). We should not log it at any
level, since it is mostly an useless information.

What we _do_ care about is invalid setting for the mentor list,
which can be done by a wiki setting their list to "Mentor|list". In that
case, WikiConfigException should be thrown (and logged), so the Growth
team can look and act as needed.

This patch intentionally keeps WikiConfigException thrown from
MentorManager logged at INFO level. Once this patch is deployed,
we can review messages that get logged and raise the log level
as appropriate.

A follow-up patch should refactor getRandomAutoAssignedMentor to not
throw WikiConfigException, and review/fix calls of getMentorForUser()
to check for null instead of WikiConfigException.

Bug: T290371
Change-Id: Iee52a0ece7fdee595b80b225fb2cfaca0175cf2b