Update suggested edits module state from start editing dialog
The server-side rendered suggested edits module HTML was appended to the
start editing module after completion of start editing. This caused a
problem in that the HTML there wouldn't necessarily be consistent with
what the user experienced in the start editing dialog; if the user
changes topics/task types, the HTML from the server-side render is
outdated. This is particularly jarring when one selects topics that
results in no results.

To work around this, broadcast updateMatchCount events from the
StartEditingDialog to the Suggested Edits module, which is now
initialized by default on desktop but is hidden to the end user. Then,
when activation of the start editing dialog is completed, unhide the
Suggested Edits module.

Bug: T290212
Change-Id: I170fa80886c7e0ec6dbd14ddf7cd899a5265c900