Suggested Edits: add timing data for time to interactive
We currently record suggestedEditsLoadingComplete, which happens after
the JS for Suggested Edits loads but also after the API request to fetch
a set of tasks completes.

Arguably the time to interactive [TTI](https://web.dev/interactive) for
the module is after the controls are functional and the first task card
has appeared; the module can be interacted with at this point (e.g. if
the API query to fetch tasks is taking a long time, the user can still
update the filters and choose a different set of tasks).

In order to do this, we need to update the controls before the API
request is issued, and to do that we need to pretend that we already
have a task queue, so we use the first task data and set the position to
zero to achieve that.

Misc:

* Remove legacy code that unhides the filters, since they're always
rendered on the server side now.

Bug: T257371
Change-Id: I97d2b3af1b2ce21c62bda0e3b41ab68a9d7ae422