QuestionPoster: Handle questions asked on files/categories
With the previous code, questions asked on a file page/category page
had a broken header: instead of linking to the page, they included
the file/category in question, which is an unexpected result.

This fixes it by prepending a colon to the link if that's the case.
Wikitext parser transfers [[:File:Foo.jpg]] into a proper-looking link;
same also applies for categories.

Bug: T289932
Change-Id: Ibcb37606b6a7c40080b5a3f191c71c0ff928c53d