updateMenteeData: Fix a typo
This is supposed to be an array_map, not array_filter. No exceptions
were thrown, as PHP corrected the int to a true value, and array_diff
didn't have any issues with comparing ints and strings.

Also no visible bug was present -- the table just contained more data
than it was supposed to.

Bug: T291836
Change-Id: I1946a5a07f0a077248d2b228ad0670de3df2a7ae