updateMenteeData: Collect more profiling data
To get more insight about performance of UncachedMenteeOverviewDataProvider,
collect more profiling data there.

The data can be sent to statsd (when --statsd is passed)
or printed to the console (when --verbose is passed).

This also switches per-wiki collecting to per-cluster,
as the script now runs in parallel (so total runtime is
no longer an interesting information).

Bug: T290609
Change-Id: I905cd469fab5f5d93ef13048413cb0783d1e8c7f