Fix JSDoc warnings
Fix most JSDoc warnings as finding CI errors in the sea of warnings
was getting cumbersome. Leaves the @note warnings (not sure how to
add the notes otherwise) and a few line length warnings where there
was no obvious way to fix due to lots of indentation and/or long
words.

The common problem of comments with message keys being too long is
tracked in wikimedia/eslint-config-wikimedia#399.

Change-Id: I31e39e081f0bd6616458415e8373e8423d567593