SuggestedEdits: Always load help panel when doing a newcomer task
What:

Add a new query parameter to the suggested edit card, 'gesuggestededit',
that is not removed from the query upon arriving at the article.

HelpPanel#shouldShowHelpPanel() can then consult this query parameter
and load the help panel, even if the user preference is disabled, as
without it, unstructured and structured tasks are incomplete.

Why:

It's easy to forget to check the box to enable help panel as a user
evaluating the feature on a community wiki or as a developer with a
fresh install using user ID 1, which was created before
GrowthExperiments onLocalUserCreated had a chance to force enable the
help panel preference.

Bug: T284088
Change-Id: I0f03278a4b827880241ac59dbe2e1df9f96b4db5