QuestionPoster: set minimum dialog height
In contexts where the home panel is shown by default, basing the dialog height on the home panel makes more sense (so the dialog height doesn't change when the panel changes). However, in the context of QuestionPoster, the home panel is never shown so the min height should be set in case the home panel's height is shorter than that of askHelp/questioncomplete.

Bug: T293262
Change-Id: I89ab4fc948cc74bc7bad981b82bb64b1e92e08f3