Add ApiMentorDashboardUpdateData
In a future commit, this will be used from
mentor dashboard's frontend, allowing the mentor
to schedule updates of the mentee overview module.

The API is tightly rate limited: we only let mentors
to request one update each 2 hours. This limit cannot
be bypassed even by admins, as the update is potentially
expensive.

Bug: T293454
Change-Id: Ia4862f8f38282d7a932142d7b77e0135c9ff5fe9