i18n: Links to translatable MW.org pages should use Special:MyLanguage
Special:MyLanguage will redirect readers to
whatever page is "best" from them language-side.

We should use Special:MyLanguage to attempt to
route the readers to a page written in a language they
speak in (the purpose of translating them), rather than
routing them to an English-speaking page (relying
on them to find the correct language in the box).

Change-Id: I6bce9452a1e243663353f8740a90f851ef49c40b