Search: Add TemplateCollectionFeature
Add a TemplateCollectionFeature for CirrusSearch that mimics
CirrusSearch's HasTemplateFeature.

The difference is that the template list is taken from the
GEInfoboxTemplates config variable instead of user input in the
search. The search can be performed with
`hastemplatecollection:infobox`, or inverted with
`-hastemplatecollection:infobox`. Other collections (for e.g. copyedit,
references, etc) can be added later.

Bug: T292141
Change-Id: I82543ce93632600049a289be025a9d7d0a94645e