Mentee overview: Add link to Special:Contributions
Specified by Figma specs, and kindly (re)requested
by cswiki community.

Omitting questions for now, as it doesn't look
Special:Contributions supports filtering for
multiple tags. The deeper level (ContribsPager
and ChangeTags::modifyDisplayQuery) look to be
multiple tags-aware, so fixing this in core
should not be a big issue?

Let's add the easy-to-add links for now,
and discuss how to do questions later.

Bug: T294230
Change-Id: I1262eafd12434cbe60e340297f4c795b25187a94