Un-community-config $wgGEHelpPanelSuggestedEditsPreferredEditor
$wgGEHelpPanelSuggestedEditsPreferredEditor assigns a default
editor to each task type handler class. Structured handlers will
break if the assigned editor is not machineSuggestions, a custom
editor that's uncomprehensible to wiki users. They will also
break if the relevant task type is not listed, which complicates
deploying new structured task types. For unstructured task types,
this configuration option would only allow all or none of them
to be defaulted to VE/wikitext, which is not useful (a per-tasktype
setting would make more sense).

Change-Id: I7e6f913c3195b8f8ae3d618de7f489558e047464