Use cached image recommendations
What:

* Introduce a cache-backed provider for accessing image recommendation
data for a title.

Why:

* The uncached provider can take enough time to fetch data that it
impacts the user experience.
* Fetching from WANObject cache is much faster and removes a significant
chunk of overhead from the response time (see for example this request
https://w.wiki/4LJb and how the call to ::get is 430ms, 31% of the wall
time)

How:

* Replace the existing image recommendation provider with a cache-backed
version that checks WANObject cache, and if nothing's found, calls the uncached
service image recommendation provider
* Add a new class, a "TaskSetListener" (better name/structure wanted)
that runs after a TaskSet is fetched in CacheDecorator
* In TaskSetListener, create a deferred update that iterates over tasks
in the taskset and fetches recommendation data for the task, caching it
for a week. (Could probably cache for longer, as the recommendation data
doesn't change much if at all currently.) When the user clicks through
to a task, we should be able to load data from the cache instead of via
HTTP request to the service.

Bug: T295032
Change-Id: Ie8e5a2c4b41f7d0321b26ccf08e56104d9650058