Quality gate: Daily limit for image recommendations
PostEditDialog:

 * Don't show image tasks if over daily limit
 * Adjust appearance when limit exceeded, and hide image-recommendation
   tasks from task fetch

Quality gate

* Allow task types to inform the frontend about JS validation that
should be performed on click
* If a task type provides validation, require and set up binding to
click events on task card and edit button.
* Add a QualityGate decorator on top of CacheDecorator, responsible for
adding in additional metadata to the TaskSet about any quality gates for
the given TaskSet.

Follow-ups:

* Post edit dialog doesn't show on desktop if a suggestion is rejected;
that's not broken by this patch, so it doesn't need to be fixed here either
* Quality gate IDs could probably be defined alongside other task type
configuration settings, and managed via Special:EditGrowthConfig

Bug: T293316
Change-Id: I063763f1ec359b1e974e62b1d74ab58de7c3bfcc