TaskSet: Add ImageRecommendationFilter
Allow for filtering out image recommendations that don't have
corresponding cache entries.

When AddImageSubmissionHandler schedules jobs to purge items from the
search index, it also adds that item to an "invalidated items" cache
Then on Special:Homepage and ApiQueryGrowthTasks, tasks which have
an entry in this cache will be filtered out from the user's view.

The TTL is set conservatively to 10 minutes. Updating a weighted
flag (unlike creating a new one) happens within a minute or two.

Bug: T295410
Change-Id: Ieeaa477e410af050f58738f387cc620e4b49cfad