Suggested edits: Make dependencies internal for TaskTypesAbFilter.js
TaskTypes.json and DefaultTaskTypes.json are only used by
TaskTypesAbFilter.js; make them internal. This simplifies the
code and makes it possible to use the task types in
convertTaskTypes() and getDefaultTaskTypes() as well, which will be
useful later.

Rename filterTaskTypes and filterDefaultTaskTypes to better reflect
their new usage pattern.

Also replace tasktype/variant names with constants, which makes
mixing up the unfortunately similar-but-not-quite-the-same task and
variant name harder.

Change-Id: Ida4ba74248014866d8354de079e2d722e09d72d3