AddImage: Refresh user's task feed after undecided rejection
If the user rejects a task with the reasons of 'unfamiliar' or
'foreignlanguage', the task is not removed from the pool of
image recommendation tasks available to all users. That
means the user could possibly encounter it in their suggested
edits feed again.

To work around that, reset the user's TaskSet cache to reduce the
likilhood that the user will encounter this task again.

Bug: T296491
Change-Id: I68a3d61b9b3950a4f5c397387e73e41979cd9f00
(cherry picked from commit d815ed63b4abbfba3d2b34b94bb3bca65054f2c1)