Community configuration: Handle disabled task types
When a task type is disabled via site configuration, hide it on
the special page and do not set it in initWikiConfig.php.

When a task type is enabled on the backend but disabled on the
frontend (this is only possible for Add Link), show a notice
on the special page but allow it to be configured.

Bug: T294518
Change-Id: Ibaf9267e832f84bb5463bde407171ebfbabe5cda