Newcomer tasks: Fix filtering of non-existent task types
Filtering should be applied after converting so a non-existent
'links' task type that got converted from 'link-recommendation'
can appropriately be filtered out.

Also replace null check with emptiness check. In theory it should
not possible for a user to have an empty filter set but if it
happens somehow, it will cause errors and replacing it with the
defaults seems like the easiest thing to do.

Bug: T296366
Change-Id: I6a046667d19a2b27ed31fa75c4dc90f2c59e539f