Maintenance script to fix structured edit tags
Use the associated log entries (which we expect to be reliable)
to tell whether something that's tagged as an Add Link or Add Image
edit really is one, and remove the tag if it isn't.

For now make no attempt to fix things in the other direction (add
the tag if there's an associated log entry) as that's a less likely
failure mode.

Example query results and explain on cswiki: P18029
Alternative query with rcid-based pagination: P18030
We go with the first here - it's less performant, but there aren't
enough structured edits for that to be a big problem, and it makes
it easier to limit updates to a given time period.

Bug: T296818
Change-Id: Ibc6e144a57646bbb12bbf49bbfaf09ab8f59bdc5