Add Image: Use tasktype, not quality gate, to pass caption limit
Setting minimumCaptionCharacterLength as a quality gate breaks
quality gate semantics: a certain quality gate type not being
present should not result in errors, but
ceRecommendedImageCaptionNode had a hardcoded dependency on
quality gate data. That's a problem because in some development
setups the quality gate decorator might be omitted.
Pass it via the task type config instead.

Change-Id: I7fab677609d380fe3aadb67ebc715916d3db4a2b