Hardcode EditGrowthConfig requiresWrite/doesWrites
These methods are called before $userCanWrite is set. In any case,
there is little point in them trying to reflect the current user's
ability to write. Instead, make doesWrites always return true
(that only means the page might make writes, not that it definitely
makes any in the current request) and make requiresWrite always
false since we still want the page to be viewable during readonly
mode (whether or not the user would be allowed to make writes).
Instead, do a readonly check on submit.

Bug: T297857
Change-Id: I790281cfc25b1837d11bf0b7a590591f66be8ba8