Use content language for mentor intro text
Mentor intro text is parsed from a wiki page written in the
content language. The various transformations that happen to it
should also use content language. Conveniently, this lets us
avoid fetching the user language in the MentorManager service
factory which would cause problems as the service might be needed
early on in the request lifecycle.

Also switch the default mentor message to content language. That's
more debatable but it makes sense to have the default and custom
messages be in the same language.

Bug: T297855
Change-Id: Iff25b088d17d2619a8a030bdbd3ab44828ef4d94