Rewrite MentorManager caching
This patch does four things:

    1. Moves caching to MentorPageMentorManager
    2. Replaces ObjectCache::getLocalClusterInstance with a WAN cache
    3. Renames the property used to store cache to wanCache
    4. Removes the TTL property

1. is done for practical purposes only, as the only other
provider inheriting from MentorManager is StaticMentorManager,
which really doesn't need any caching -- and I didn't want to
make caching optional (by not requiring user to pass any WAN cache)
just to keep it in the abstract class.

2. is done as suggested in discussion to I1100051. The local-cluster
cache does not really have any benefits when compared to the WAN cache,
so there is no reason to use it in MentorManager's code.

3. is done in anticipation of a different kind of cache being
added (an APCu cache) to MentorManager. Using a specific name
will make it easier for us to tell both kinds of cache apart.

4. is done to follow the usual cache access method in rest of GE code
(where it uses a WAN cache passed as __construct's parameter and not
a setCache method, like MentorManager did prior to this patch).

Bug: T298048
Change-Id: I8aa11c59a42a65e2874ee412807176553d9e5dfc