QualityGates: Enforce daily limit on server-side
Without this, users can open tasks in a new tab and bypass the
client-side validation. Normally that isn't a big deal, but it's
possible some users are doing that to complete a larger number of tasks.

It would be nice to not have this logic here, but not sure we have a
better place for it at the moment.

Bug: T274325
Change-Id: I19550725ca002872eea3f5a94cc0356595b338ca