Move mentor pool code to MentorProvider
MentorProviders are responsible for knowing who the mentors are
as well as getting required detailed information about them (currently,
their UserIdentity and introduction message). Factor related code
out of MentorManager.

This should get merged as close after a branch cut as possible and
tested in various mentorship-related scenarios (no mentors, only one mentor, who
is a present mentor/away/the mentee as well, etc.), as not all of them
are covered by tests.

A next commit will introduce JsonMentorProvider, that will be able
to get a list of mentors from a JSON stored in a wiki page (similar
to what community configuration does).

Bug: T264343
Change-Id: I21d75c6c191bb61ffddc91fcb1e2413f7737ae66